refactor: helper functions to return credentialFunc move to own file credential.go
